add phases to unparser/started ngram impl

+++ src/scripts/run_fuzzer.py
@@ -1,12 +1,57 @@
 
 from common import *
 
-def generate_test():
-    env = Shared.Environment()
-    env.attr.expr.depth = 3
+def is_generated(env):
+    if env.attr_exists('.test.metadata.paths.dm_file'):
+        return True
+    else:
+        return False
+
+def generate_test(tenv):
+    if is_generated(tenv):
+        return
+    benv = Shared.Environment()
+    benv.attr.expr.depth = 3
     builder = DreamCollider.FullRandomBuilder( )
-    builder.generate( env )
-    return builder
+    builder.generate( benv )
+    builder.unparse(ngrams=tenv.attr.ngrams)
+    tenv.attr.collider.builder = builder
+
+def try_init_test_instance(env):
+    env.attr.test.metadata.name = Shared.Random.generate_string(24)
+    env.attr.test.root_dir = env.attr.tests.root_dir / env.attr.test.metadata.name
+    DMTR.Metadata.load_test(env)
+
+def save_test_dm(tenv):
+    tenv.attr.test.metadata.paths.dm_file = 'test.dm'
+    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.dm_file, "w") as f:
+        f.write( tenv.attr.collider.builder.text )
+
+def save_test_ngrams(tenv):
+    tenv.attr.test.metadata.paths.ngram_info = 'ngram_info.json'
+    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.ngram_info, "w") as f:
+        f.write( json.dumps( tenv.attr.collider.builder.ngram_info ) )
+
+def save_collider_info(tenv):
+    tenv.attr.test.metadata.paths.collider_ast = 'collider_ast.txt'
+    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.collider_ast, "w") as f:
+        s = io.StringIO()
+        tenv.attr.collider.builder.print(s)
+        f.write( s.getvalue() )
+
+    tenv.attr.test.metadata.paths.collider_model = 'collider_model.json'
+    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.collider_model, "w") as f:
+        f.write( json.dumps(tenv.attr.collider.builder.get_model(), cls=DreamCollider.ModelEncoder) )
+
+async def run_test(env):
+    ctenv = env.merge( baseenv.attr.envs.byond )
+
+    await DMTR.Runners.byond_codetree(ctenv)
+    await DMTR.Runners.opendream_ast(ctenv)
+    await DMTR.Runners.clopen_ast(ctenv)
+    #await DMTR.Runners.run_meta(ctenv)
+
+    return ctenv
 
 async def print_main():
     builder = generate_test()
@@ -23,45 +68,6 @@ async def print_many_main():
         if random.random() < 0.001:
             print("====================================")
             print( builder.unparse() )
-
-def try_init_test_instance(env):
-    env.attr.test.metadata.name = Shared.Random.generate_string(24)
-    env.attr.test.root_dir = env.attr.tests.root_dir / env.attr.test.metadata.name
-    DMTR.Metadata.load_test(env)
-
-def is_generated(env):
-    if env.attr_exists('.test.metadata.paths.dm_file'):
-        return True
-    else:
-        return False
-
-def generate_test_and_save(tenv):
-    if is_generated(tenv):
-        return
-    builder = generate_test()
-    tenv.attr.test.metadata.paths.dm_file = 'test.dm'
-    tenv.attr.test.metadata.paths.collider_ast = 'collider_ast.txt'
-    tenv.attr.test.metadata.paths.collider_model = 'collider_model.json'
-    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.dm_file, "w") as f:
-        source = builder.unparse()
-        f.write( source )
-    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.collider_ast, "w") as f:
-        s = io.StringIO()
-        builder.print(s)
-        f.write( s.getvalue() )
-    with open( tenv.attr.test.root_dir / tenv.attr.test.metadata.paths.collider_model, "w") as f:
-        f.write( json.dumps(builder.get_model(), cls=DreamCollider.ModelEncoder) )
-    DMTR.Metadata.save_test(tenv)
-
-async def run_test(env):
-    ctenv = env.merge( baseenv.attr.envs.byond )
-
-    await DMTR.Runners.byond_codetree(ctenv)
-    await DMTR.Runners.opendream_ast(ctenv)
-    await DMTR.Runners.clopen_ast(ctenv)
-    #await DMTR.Runners.run_meta(ctenv)
-
-    return ctenv
 
 async def find_new_errors(path, tmp_path):
     env = Shared.Environment()
@@ -81,7 +87,10 @@ async def find_new_errors(path, tmp_path):
     while True:
         tenv = env.branch().merge(empenv)
         try_init_test_instance(tenv)
-        generate_test_and_save(tenv)
+        generate_test(tenv)
+        save_test_dm(tenv)
+        save_collider_info(tenv)
+        DMTR.Metadata.save_test(tenv)
         if is_generated(tenv):
             await run_test(tenv)
 
@@ -135,8 +144,9 @@ async def generate_batch(output_dir, path, n, *args):
     env = Shared.Environment()
     env.attr.tests.root_dir = Shared.Path( path )
     env.attr.tests.required_test_count = int(n)
+    env.attr.ngrams = DreamCollider.NGram()
 
-    if "reset" in args:
+    if "reset" in args and os.path.exists( env.attr.tests.root_dir ):
         shutil.rmtree( env.attr.tests.root_dir )
     # count existing tests
     exist_test_count = 0
@@ -153,7 +163,11 @@ async def generate_batch(output_dir, path, n, *args):
     while current_test_count < env.attr.tests.required_test_count:
         tenv = env.branch()
         try_init_test_instance(tenv)
-        generate_test_and_save(tenv)
+        generate_test(tenv)
+        save_test_dm(tenv)
+        save_test_ngrams(tenv)
+        #save_collider_info(tenv)
+        DMTR.Metadata.save_test(tenv)
         current_test_count += 1
         generated_test_count += 1
     print( f"generated {generated_test_count} tests")
